improved docs for agent state and zome api functions
