+++ core/src/persister.rs
@@ -52,7 +52,7 @@ mod tests {
 
         let state = State::new();
 
-        let action = ::state::Action::Agent(::agent::Action::Commit {
+        let action = ::state::Action::Agent(::agent::state::Action::Commit {
             entry: test_entry(),
             id: snowflake::ProcessUniqueId::new(),
         });
